+++ run_tests.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
-poetry run pytest --cov=src/ --cov-report html
+poetry run pytest --cov=src/ --cov-report html $1
 poetry run mypy src/